requirements.txt
streamlit
pandas
numpy
scikit-learn
requests
python-dotenv
joblib
geopy
sqlite3